Update: migrate to decimal type in TransactionTariffEntity class
Refs: (#2)

+++ database/07-accounting.sql
@@ -70,7 +70,7 @@ create table if not exists Accounting.Transaction_Tariff
 (
     transactionId varchar(15) not null,
     tariffId int not null,
-    amount float not null,
+    amount decimal(18, 2) not null,
     primary key (transactionId, tariffId),
     foreign key (transactionId) references Accounting.Transaction,
     foreign key (tariffId) references Accounting.Tariff
